Finish writing data into database
Note: data fetched from sheets should be all strings for now since
the correct data format can't be guaranteed on the sheet side! This
can be improved in the future by telling the sheet owners to pay
attention when filling out a sheet cell and use data validations on
the sheet.